reorganise EditableBooster and StringBooster to and from conversion method
